Finishing PuzzleActivity2.kt, still have to sort out lifecycle problems which affect pause fragment and timer management. Also have to add new local database functionality for saving data using Room. Unit Testing also must be updated, need to mock Android components properly and update the 2nd half of tests to changed functions.

+++ app/src/main/java/com/example/lamelameo/picturepuzzle/PuzzleActivity.java
@@ -650,11 +650,11 @@ public class PuzzleActivity extends AppCompatActivity implements PauseMenu.OnFra
         // for each row loop 4 times creating a new cropped image from original bitmap and add to adapter dataset
         for (int x = 0; x < columns; x++) {
             // for each row, increment y value to start the bitmap at
-            float ypos = x * cellSize;
+            int ypos = x * cellSize;
             for (int y = 0; y < rows; y++) {
                 // loop through 4 positions in row incrementing the x value to start bitmap at
-                float xpos = y * cellSize;
-                Bitmap gridImage = Bitmap.createBitmap(bmp, (int) xpos, (int) ypos, cellSize, cellSize);
+                int xpos = y * cellSize;
+                Bitmap gridImage = Bitmap.createBitmap(bmp, xpos, ypos, cellSize, cellSize);
                 // converted to drawable for use of setImageDrawable to easily swap cell images
                 bitmaps.add(new BitmapDrawable(getResources(), gridImage));
             }
